Make sure .item returns a read-only container
Issue #5908 showed that the container that .item should return,
should either be the invocant (if it is a container already), or
a read-only container if it is not.

This commit achieves that by introducing a helper sub called
"read-only-container" that will return its argument (which is
is a read-only container because of the way untyped positional
arguments are handled).

This is arguably a bit of a hack, but an nqp:: based solution
escapes me as of now.  At least this will now be fixed for the
2025.06 release.