Make sure coercion don't introduce a container
Commit f95f85a made sure that Mu.item returns a read-only container
if the invocant wasn't already container (as opposed to a right
value before that).

This somehow caused a test failure in the PDF distribution (see #5916).

The failing test needed to have the following prerequisites:
- a role that has a COERCE method
- a subset of that role
- a variable with a constraint of that role
- an enum with a string value that matches the subset's where condition
- assignment of that variable using the enum

so:

  role A { method COERCE($str) { $str but A }; method foo() { ... }
  subset B of A where "A" | "B" | "C";
  enum C << :C<C> >>;
  my B() $d;
  $d = C;
  $d.foo;  # fails
  No such method 'foo' for string 'C'. Possible causes are:
    - You actually called 'foo' on a container, was that what you
    intended?
    - Found 'foo' on type 'A'

Note that all of these just worked:

  $d = "C" but A;
  $d = A.COERCE("C");

so it wasn't the COERCE logic or the "but" infix doing weird things.

After investigation (by putting a "die" in the COERCE logic, and running
the test file with --ll-exception), it turned out that line 4447 in
src/vm/moar/dispatchers.nqp:

  (my $coerced_value := $method($nominal_target, $value))

actually returned a container (when it obviously?? shouldn't).

This commit adds an extra nqp::decont() in that logic, and the similar
logic just above.  This doesn't break any spectests.  But it does feel
like a stopgap.  At least this won't block the 2025.06 release now.