Make Enumeration:D.clone return invocant
Because instantiated Enumerations are value types, a .clone should
return the invocant unaltered, as there can only be one of a value
type.

Because Enumeration:U would also need to return the invocant, take
any Enumeration: as invocant.

Because default attribute values in a role are cloned (because a
role doesn't really exist as code until it is consumed), this code:

  my enum E <one>;
  role A { has $.a = one }
  say one =:= A.new.a;

would return False, rather than True.  This also affected the infix
=== operator in a similar way, because that assumes =:= semantics
under the assumption that there is always only one of a value type.

Fixes #5599