Fix program name in USAGE
With the new script wrappers the script is not executable anymore. With
the new wrappers we always generate two files, a wrapper program that is
executable and a .raku script that is not executable. Thus the path
shortening logic isn't triggered anymore. Fix by additionally detecting
such wrapper + script combinations and shortening them respectively as
well.

There is a really tiny security concern, because we technically can't know
if that wrapper really was used to call the script. But given it's only
used to display the scripts name in USAGE information text, I believe this
is a risk we can confidently take.