Make pick/roll(iterable) return a sensible value
Before, it would also be returning an empty list, because of the
lack of a second argument (the set to choose from).

This now checks for the first element being an iterable if there
is nothing valid as a second argument, and then interpretes that
as a pick/roll of a single element on that iterable.

This makes pick(11..20) return a random number between 11 and 20
inclusive.  Same with roll(11..20).

Spotted by msiism++
  https://irclogs.raku.org/raku-beginner/2025-07-17.html#19:57