Add exit-ok to Test module
As promised in https://github.com/Raku/problem-solving/issues/492#issuecomment-3263735792
the "exit-ok" subroutine takes a Callable (which is expected to call
exit() ), an integer value (0 by default), and an optional description.

If the code performs an "exit", then the value will be checked: it
will be judged ok if the value matches, else it will be a nok.

If the code does not execute an "exit", then it will also be a nok.