v6.e - Attributes with where clause must have initializers...
... or the `is required` trait.

Issue #5915 shared this expression:

```
use v6.*; my class A { has $.a where * > 0 }
```

The where clause is run during the composition of the
declaring class. In base this produced a SORRY about no available
candidates for the `>` operator and in AST this produced warnings
about Any. In both cases the compiler was complaining about an
undefined value being used in numeric context.

This is because the attribute `$!a` has no initializer and is not
marked as `required` through the `is required` trait, thus
`$!a` in v6.d could be in a messed up kind state if anyone forgot
to pass a value in the constructor.