RakuAST/v6.e - Fix subset type defaults in containers
The proper place for maybe-nominalization (which in turn also
involves maybe-definite-ization, for versions < v6.e.

Only RakuAST needs to be fixed here, thankfully, as I have had
way too high a does of metamodel internals for the moment.

Essentially this change is required because subsets are technically
nominalizable, but the call to nominalizable returns the refinee
($!of) or the nominalized refinee.

But the base-case (ie, before these nominalizable checks throughout
Perl6::Metamodel classes was added) returned the subset type object
directly. (Subsets are apparently always nominalizable, as the code
attests and my experiments confirmed).

In order to return that state of affairs, we just call
IMPL-MAYBE-NOMINALIZE as we generate type metaobjects.

R#5912 (#5912)