RakuAST: fix =table checks for numbered tables
The problem was that the $*IN_TABLE check was done on the whole
type, instead of the root of the type.  So =begin table3 would
check for "table" and see "table3" and not increment $*IN_TABLE.

Now the root of doc-identifier is also exposed, and used to check
whether or not we're in a table.

Fixes #5994