RakuAST: allow ::Block.new() to take :$multiness
This is apparently being passed to it in some situations, but does
not apparently serve any purpose in the ::Block case.

This may be a band-aid type of fix, but since there is apparently
already the same situation for :$signature, it feels it is ok to
apply a similar fix for this case.

Fixes #5997