Prevent issues when multiple threads try to load the same precomp
Until now, we were only locking the hash of already `$loaded` precomp files for
the moment where we check if something is in it for the slot we are interested
in, and then again when we want to put something new in.

However, we were not preventing the situation where multiple threads check
at the same time, all get the answer "nothing there yet", and all going ahead
to load the precomped file, and then setting the result in the hash regardless
of a value that may already have appeared in the mean time.

Another approach to this would be simply to not install a result after loading
the precomp file if another value has already appeared in the mean time, and
throwing the redundant loaded bytecode files away. However, this means that
threads will do unnecessary work if we're unlucky.

The approach in this commit installs a Promise in the `$loaded` hash if there
was nothing there yet, and awaits the result if there was a Promise at that
time. When the precompilation file is loaded, the promise is kept (for any
waiting threads) and the slot is replaced with the result (so that we don't
pay the overhead of Promises after this). If there was an issue that means
the thread that "owns" the Promise can't fulfill it, it breaks the Promise
and frees the slot in the hash. Waiting threads will then try again, meaning
they will either install a Promise of their own if applicable, or get a
"nothing here" result.

This seems to fix #1920 and offers a quite fine locking granularity.