Reduce size of generated rakuast source code files
These files are put into the BOOTSTRAP/v6c.nqp file and contribute a
good amount of lines to it.

One thing this commit does is remove the decont of arguments, including
the `$SELF` argument, if they are not mentioned in the body of a method.

Another thing it does is cut off whitespace from the front and back of
a method's body, and changes how the first line is indented. This has
the actual effect of removing a lot of lines that are empty except for
whitespace.

The number of lines in gen/moar/ast.nqp goes from 41845 to 36313 (a
reduction of 13%) and the number of bytes goes from 2142295 down to
2065410 (a reduction of 3.5%).

It's not going to make an enormous difference in terms of performance,
or reduce the parse time, especially since whitespace should be pretty
quick to skip over, but maybe after these lowest of hanging fruit
there's more to be had in here ...