RakuAST: don't allow :{ } if it's a block
Addresses https://github.com/rakudo/rakudo/issues/1559 in RakuAST.

This removes the RakuAST::Block.block-or-hash method, and re-instates
that as a "block-or-hash" sub in the Actions (where it should have
been in the first place, as this has nothing to do with artificially
building ASTs, but purely with the Raku grammar).

Rebuilding that logic also allowed an easy way to indicate that an
attempt was made to create a block with the :{ } syntax, and act
appropriately with a panic.