coerce $pos to Int in the .contains multi for a Regex $needle
Str.contains() requires the second positional to be an Int. So one needs
to convert the Cool $pos to Int, just as in the three multis above.
This should fix the fact that Rakudo freezes when calling .contains
with a Regex for $needle and non-Int Cool for $pos, e.g.:
	12345.contains(/5/, "1")
and probably also
	"abc".contains(/c/, "1")