Normalize rakudo-home path
The script wrappers generate paths with many updirs (`..`). When the CURI
is configured to use relative localization of the `rakudo` executable, they
take the script executable path

`some/dir/share/perl6/site/bin/zef`

and then navigate up to the `rakudo` executable, resulting in:

`some/dir/share/perl6/site/bin/../../../../bin/rakudo`

The resulting `rakudo-home` then is

`some/dir/share/perl6/site/bin/../../../../share/perl6`

This `rakudo-home` is then the basis for all paths dealing with the
standard CURIs. That path navigating in, out and then back in is mildly
irritating.

Fix this by normalizing the `rakudo-home` path.

The implemented logic does not do fancy resolving of symlinks and similar.
This is surely not perfect but will go a long way.