[RakuAST] Fix parameterizations of types in `is` trait
Considering

    role R[::T] { has @.a is Array[T] }
    my $r = R[Int].new: :a(1,2,3)
    # Illegal binding exception, "expected `T` and got `Int`"

This would previously fail because the type capture `T`
was "frozen" into the parameterization of `@.a`.

This patch fixes the above issue, which only affected
types used with the `is` trait.

Please see comments inline for more details, but essentially what
we do is add some significant fiddling around when a parameterized
type is passed to `is`. We do this by setting an explicit container
base type and also setting a specialized `$!key-type` attribute to
`RakuAST::VarDeclaration::Simple` that is then used in parameterizations.

Note that `RakuAST::VarDeclaration::Simple` is probably due for a bit of
untangling, but it can wait for now.

R#5970 (#5970)