Make -MSIL cleaner by excluding any core BBs
Adds a :skip-core named argument to the SIL subroutine that will
execute a "use Test" and record the BBs seen from that.  Those
BBs will then not be reported.

The logic behind executing "use Test" is to execute at least part
of the grammar / actions, and the module loading logic.

A piece of code such as "my $a = 42; +$a for ^1000" will now clearly
show that prefix:<+> and the Numeric method get inlined:

Spesh Inline Log Report of Process #69418 (2025-12-27T13:03:20Z)
Executing: -e my $a = 42; +$a for ^1000

Successful inlines
--------------------------------------------------------------------------------
    BB(1, 96 bytes) -> <unit> BB(2)
    Numeric BB(3462, 10 bytes) -> prefix:<+> BB(3487)
    prefix:<+> BB(3487, 100 bytes) -> BB(1)
    BB(3647, 46 bytes) -> <unit> BB(2)
--------------------------------------------------------------------------------