Simplify handing of RAKUDO_OPT environment variable
Instead of making RAKUDO_OPT provide defaults for the command-line
arguments handling, simply splice the RAKUDO_OPT arguments into
the arguments (after the initial argument, which is the executable
to be used).

This fixes the issue described in #6057, but at the expense of
subtly changing the semantics of RAKUDO_OPT (although I doubt that
anybody actually was using -Mmodule in RAKUDO_OPT before, as this
is what this research instigated).

This also fixes the behaviour as reported by ugexe++ in
  https://github.com/rakudo/rakudo/issues/6057#issuecomment-3694259810