Make sure RAKUDO_OPT doesn't get repeated when precomping
Any options of --ll-exception, --profile, --optimize and --target
are passed to the process doing precompilation.  If any of these
was specified in the RAKUDO_OPT, they would be passed a second
time (and a third when recursing, etc, etc).

So it's important to remove RAKUDO_OPT from the environment before
recursing into another instantiation of rakudo.

Apparently it is ok to have specified the above arguments (apparently
for debugging precompilation issues), so at this point it doesn't
matter whether these arguments originated from RAKUDO_OPT, or where
actually specified on the command line.